refactor(webtests): share API token insert plumbing and drop adminBearerReq
testingRequest already builds bearer requests; insertAPITokenRow serves both
the sha256 and the legacy pbkdf2 seed helpers.

address pr-swarm finding: conventions + design
