+++ pkg/webtests/admin_test.go
@@ -19,9 +19,7 @@ package webtests
 import (
 	"net/http"
 	"net/http/httptest"
-	"strings"
 	"testing"
-	"time"
 
 	"code.vikunja.io/api/pkg/config"
 	"code.vikunja.io/api/pkg/db"
@@ -30,7 +28,6 @@ import (
 	"code.vikunja.io/api/pkg/models"
 	"code.vikunja.io/api/pkg/modules/auth"
 	"code.vikunja.io/api/pkg/user"
-	"code.vikunja.io/api/pkg/utils"
 
 	"github.com/labstack/echo/v5"
 	"github.com/stretchr/testify/assert"
@@ -58,44 +55,9 @@ func adminReq(t *testing.T, e *echo.Echo, method, path string, u *user.User, bod
 	if u != nil {
 		tok, err := auth.NewUserJWTAuthtoken(u, "test-session-id")
 		require.NoError(t, err)
-		bearer = tok
+		bearer = "Bearer " + tok
 	}
-	return adminBearerReq(e, method, path, bearer, body)
-}
-
-func adminBearerReq(e *echo.Echo, method, path, bearer, body string) *httptest.ResponseRecorder {
-	req := httptest.NewRequest(method, path, strings.NewReader(body))
-	req.Header.Set(echo.HeaderContentType, echo.MIMEApplicationJSON)
-	if bearer != "" {
-		req.Header.Set(echo.HeaderAuthorization, "Bearer "+bearer)
-	}
-	res := httptest.NewRecorder()
-	e.ServeHTTP(res, req)
-	return res
-}
-
-// insertAPIToken writes the row directly, bypassing Create's permission
-// validation so legacy scope keys can be seeded. Returns the cleartext token.
-func insertAPIToken(t *testing.T, ownerID int64, perms models.APIPermissions) string {
-	t.Helper()
-
-	cleartext, err := utils.CryptoRandomString(40)
-	require.NoError(t, err)
-	cleartext = models.APITokenPrefix + cleartext
-	token := &models.APIToken{
-		Title:          "admin scope test token",
-		TokenSha256:    models.HashAPIToken(cleartext),
-		APIPermissions: perms,
-		ExpiresAt:      time.Now().Add(24 * time.Hour),
-		OwnerID:        ownerID,
-	}
-
-	s := db.NewSession()
-	defer s.Close()
-	_, err = s.Nullable("token_salt", "token_hash", "token_last_eight").Insert(token)
-	require.NoError(t, err)
-	require.NoError(t, s.Commit())
-	return cleartext
+	return testingRequest(e, method, path, body, bearer)
 }
 
 func TestAdmin_APIToken(t *testing.T) {
@@ -108,31 +70,31 @@ func TestAdmin_APIToken(t *testing.T) {
 
 	t.Run("named scope reaches a PATCH route", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_set_status"}})
-		res := adminBearerReq(e, http.MethodPatch, "/api/v1/admin/users/2/status", tok, `{"status":0}`)
+		res := testingRequest(e, http.MethodPatch, "/api/v1/admin/users/2/status", `{"status":0}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 	})
 
 	t.Run("legacy scope key still authorises", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_status"}})
-		res := adminBearerReq(e, http.MethodPatch, "/api/v1/admin/users/2/status", tok, `{"status":0}`)
+		res := testingRequest(e, http.MethodPatch, "/api/v1/admin/users/2/status", `{"status":0}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 	})
 
 	t.Run("other admin scope is denied", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_list"}})
-		res := adminBearerReq(e, http.MethodPatch, "/api/v1/admin/users/2/status", tok, `{"status":0}`)
+		res := testingRequest(e, http.MethodPatch, "/api/v1/admin/users/2/status", `{"status":0}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusUnauthorized, res.Code)
 	})
 
 	t.Run("admin-only token is denied outside admin", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_list", "users_set_status"}})
-		res := adminBearerReq(e, http.MethodGet, "/api/v1/tasks/all", tok, "")
+		res := testingRequest(e, http.MethodGet, "/api/v1/tasks/all", "", "Bearer "+tok)
 		assert.Equal(t, http.StatusUnauthorized, res.Code)
 	})
 
 	t.Run("non-admin owner is gated", func(t *testing.T) {
 		tok := insertAPIToken(t, 2, models.APIPermissions{"admin": {"users_set_status"}})
-		res := adminBearerReq(e, http.MethodPatch, "/api/v1/admin/users/3/status", tok, `{"status":0}`)
+		res := testingRequest(e, http.MethodPatch, "/api/v1/admin/users/3/status", `{"status":0}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusNotFound, res.Code)
 	})
 }
