+++ pkg/webtests/huma_admin_actions_test.go
@@ -630,38 +630,38 @@ func TestHumaAdminAPIToken(t *testing.T) {
 	t.Run("named scopes reach the PATCH routes", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_set_admin", "users_set_status", "projects_set_owner"}})
 
-		res := adminBearerReq(e, http.MethodPatch, "/api/v2/admin/users/2/admin", tok, `{"is_admin":true}`)
+		res := testingRequest(e, http.MethodPatch, "/api/v2/admin/users/2/admin", `{"is_admin":true}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 
-		res = adminBearerReq(e, http.MethodPatch, "/api/v2/admin/users/3/status", tok, `{"status":0}`)
+		res = testingRequest(e, http.MethodPatch, "/api/v2/admin/users/3/status", `{"status":0}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 
-		res = adminBearerReq(e, http.MethodPatch, "/api/v2/admin/projects/2/owner", tok, `{"owner_id":2}`)
+		res = testingRequest(e, http.MethodPatch, "/api/v2/admin/projects/2/owner", `{"owner_id":2}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 	})
 
 	t.Run("legacy scope keys still authorise", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_admin", "users", "projects_owner"}})
 
-		res := adminBearerReq(e, http.MethodPatch, "/api/v2/admin/users/2/admin", tok, `{"is_admin":false}`)
+		res := testingRequest(e, http.MethodPatch, "/api/v2/admin/users/2/admin", `{"is_admin":false}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 
-		res = adminBearerReq(e, http.MethodGet, "/api/v2/admin/users", tok, "")
+		res = testingRequest(e, http.MethodGet, "/api/v2/admin/users", "", "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 
-		res = adminBearerReq(e, http.MethodPatch, "/api/v2/admin/projects/2/owner", tok, `{"owner_id":1}`)
+		res = testingRequest(e, http.MethodPatch, "/api/v2/admin/projects/2/owner", `{"owner_id":1}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusOK, res.Code, res.Body.String())
 	})
 
 	t.Run("other admin scope is denied", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_list"}})
-		res := adminBearerReq(e, http.MethodPatch, "/api/v2/admin/users/2/admin", tok, `{"is_admin":true}`)
+		res := testingRequest(e, http.MethodPatch, "/api/v2/admin/users/2/admin", `{"is_admin":true}`, "Bearer "+tok)
 		assert.Equal(t, http.StatusUnauthorized, res.Code)
 	})
 
 	t.Run("admin-only token is denied on tasks", func(t *testing.T) {
 		tok := insertAPIToken(t, 1, models.APIPermissions{"admin": {"users_list", "users_create", "users_set_admin"}})
-		res := adminBearerReq(e, http.MethodGet, "/api/v2/tasks", tok, "")
+		res := testingRequest(e, http.MethodGet, "/api/v2/tasks", "", "Bearer "+tok)
 		assert.Equal(t, http.StatusUnauthorized, res.Code)
 	})
